+++ lib/queries/customer.queries.ts
@@ -1,0 +1,15 @@
+export const CUSTOMER_CREATE_MUTATION = /* GraphQL */ `
+  mutation customerCreate($input: CustomerCreateInput!) {
+    customerCreate(input: $input) {
+      customer {
+        id
+        email
+      }
+      customerUserErrors {
+        code
+        field
+        message
+      }
+    }
+  }
+`;